% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA.Anal}
\alias{ANOVA.Anal}
\title{One-way ANOVA}
\usage{
ANOVA.Anal(dataSet, analSet, nonpar = FALSE, thresh = 0.05,
  post.hoc = "fisher")
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{nonpar}{If \code{FALSE} - use classical ANOVA; if \code{TRUE} - Kruskal Wallis Test}

\item{thresh}{Threshold of significance.}

\item{post.hoc}{Post-hoc statistics: \code{"tukey"} or \code{"fisher"}}
}
\value{
Native \code{analSet} with one added \code{$aov} element consisting of:
\itemize{
\item\code{$aov.nm} - method of analysis
\item\code{$raw.thresh} - value of \code{thresh} argument
\item\code{$thresh} - \code{-log10(thresh)}
\item\code{$p.value} - p-values
\item\code{$p.log} - \code{-log10(p.value)}
\item\code{$inx.imp} - logical vector of features with significant difference
\item\code{$post.hoc} - post-hoc statistics, value of \code{post.hoc} argument
\item\code{$sig.mat} - data frame of significant features with corresponding statistics
}
}
\description{
Performs one-way ANOVA. Creates a file with results:
either \code{"anova_nonparametric.csv"} or \code{"anova_posthoc.csv"}
depending on \code{nonpar} argument.
}

