% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ANOVA2.R
\name{ANOVA2.Anal}
\alias{ANOVA2.Anal}
\title{Two-way ANOVA}
\usage{
ANOVA2.Anal(dataSet, analSet, thresh = 0.05, p.cor = "fdr", type = "b")
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{thresh}{Threshold of significance.}

\item{p.cor}{Method of p-value correction, one of \code{\link[stats]{p.adjust}} methods}

\item{type}{Type of comparison. If \code{"b"} - between subjects; if \code{"w"} - within subjects}
}
\value{
Native \code{analSet} with one added \code{$aov2} element consisting of:
\itemize{
\item\code{$type} - value of \code{type} argument
\item\code{$sig.nm} - name of file with the resulting ANOVA table
\item\code{$thresh} - value of \code{thresh} argument
\item\code{$multi.c} - value of \code{p.cor} argument
\item\code{$sig.mat} - data frame of significant features
\item\code{$vennC} - Venn counts \code{\link{getVennCounts}}
}
}
\description{
Performs two-way ANOVA. Creates a file with results:
either \code{"anova_within_sbj.csv"} or \code{"anova_between_sbj.csv"}
depending on \code{type} argument.
}

