% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_EBAM.R
\name{EBAM.A0.Anal}
\alias{EBAM.A0.Anal}
\alias{EBAM.Cmpd.Anal}
\title{EBAM Fudge Factor}
\usage{
EBAM.A0.Anal(dataSet, analSet, paired = FALSE, var.equal = TRUE)

EBAM.Cmpd.Anal(dataSet, analSet, method = "z.ebam", A0 = 0,
  paired = FALSE, var.equal = TRUE)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{paired}{Are values in data set paired or not.}

\item{var.equal}{Are variances assumed equal or not.}

\item{method}{A character string or name specifying the method or function
that should be used in the computation of the expression score z.
If \code{"z.ebam"}, a modified t- or F-statistic, respectively, will be computed as proposed by Efron et al. (2001).
If \code{"wilc.ebam"}, a (standardized) Wilcoxon sum / signed rank statistic will be used as expression score.
Note: if method is \code{"wilc.ebam"}, then \code{A0} and \code{var.equal} parameters will be ignored}

\item{A0}{Fudge Factor. \code{\link{EBAM.A0.Anal}}}
}
\value{
\code{EBAM.A0.Anal} - native \code{analSet} with one added \code{$ebam.a0} element containing
standard \code{\link[siggenes]{find.a0}} function output

\code{EBAM.Cmpd.Anal} - native \code{analSet} with one added \code{$ebam} element containing
standard \code{\link[siggenes]{ebam}} function output
}
\description{
\code{EBAM.A0.Anal} - finds EBAM Fudge Factor. Uses \code{\link[siggenes]{find.a0}} function.

\code{EBAM.Cmpd.Anal} - performs Empirical Bayesian Analysis of Microarray (and Metabolites).
Uses \code{\link[siggenes]{ebam}} function.
}
\section{Functions}{
\itemize{
\item \code{EBAM.A0.Anal}: 
}}

