% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FC.R
\name{FC.Anal}
\alias{FC.Anal}
\title{Fold change}
\usage{
FC.Anal(dataSet, analSet, fcthresh = 2, percent.thresh = 0.75,
  cmpType = 0)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{fcthresh}{Fold-change threshold.}

\item{percent.thresh}{Sig. count threshold (for paired data only).}

\item{cmpType}{Comparison type. If equal \code{0} then group 1 is compared against group 2. Otherwise, vice versa.}
}
\value{
Native \code{analSet} with one added \code{$fc} element consisting of:
\itemize{
\item\code{$paired} - are data paired or not
\item\code{$raw.thresh} - value of \code{fcthresh} argument
\item\code{$max.thresh} - upper log-threshold of fold change
\item\code{$min.thresh} - lower log-threshold of fold change
\item\code{$fc.all} - fold changes of all features
\item\code{$fc.mat} - matrix of fold changes (for paired data only)
\item\code{$fc.log} - log2 of fold changes (for unpaired data only)
\item\code{$inx.up} - logical vector of increasing features (for paired data only)
\item\code{$inx.down} - logical vector of decreasing features (for paired data only)
\item\code{$inx.imp} - logical vector of features with significant difference (for unpaired data only)
\item\code{$sig.mat} - data frame of significant features
}
}
\description{
Fold change (FC) analysis is to compare the absolute value change between two group means.
Since column-wise normalization (i.e. log transformation, mean-centering) will significantly
change the absolute values, FC is calculated as the ratio between two group means using data
before column-wise normalization was applied.
}
\details{
For paired analysis, the program first counts the number of pairs with consistent change above the given FC threshold.
If this number exceeds a given count threshold, the variable will be reported as significant.
Writes an output file \code{"fold_change.csv"}.
}

