% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster_Kmeans.R
\name{Kmeans.Anal}
\alias{Kmeans.Anal}
\title{K-means analysis}
\usage{
Kmeans.Anal(dataSet, analSet, clust.num)
}
\arguments{
\item{clust.num}{The cluster number.}
}
\value{
Native \code{analSet} with one added \code{$kmeans} element containing
standard \code{\link[stats]{kmeans}} output.
}
\description{
Perform K-means analysis. Uses \code{\link[stats]{kmeans}} function.
}

