% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA.Loadings}
\alias{PCA.Loadings}
\title{PCA loadings}
\usage{
PCA.Loadings(dataSet, analSet, inx1 = 1, inx2 = 2)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{inx1,inx2}{The order numbers of PCs.}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$pca$load.x.uniq} - ???
\item\code{$pca$imp.loads} - loading matrix
}
}
\description{
PCA loadings.
Adds loading matrix to \code{analSet$pca} element.
}

