% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Correlations.R
\name{PlotCorr}
\alias{PlotCorr}
\title{Plot pattern correlations}
\usage{
PlotCorr(dataSet, analSet, imgName = "ptn_", format = "png", dpi = 72,
  width = NA)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Plot the top 25 features correlated to a feature or a pattern.
}

