% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{PlotCorrHeatMap}
\alias{PlotCorrHeatMap}
\title{Plot correlations heatmap}
\usage{
PlotCorrHeatMap(dataSet, analSet, imgName = "corr_heat_", format = "png",
  dpi = 72, width = NA, cor.method = "pearson", colors = "default",
  viewOpt = "overview", fix.col = F, no.clst = FALSE, top = F,
  top.num = 999)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{cor.method}{Method of correlation calculation, one of \code{"pearson"}, \code{"kendall"}, \code{"spearman"}}

\item{colors}{The color contrast. One of \code{"default"}, \code{"gbr"} (red/green), \code{"heat"}, \code{"topo"}, \code{"gray"}}

\item{viewOpt}{View mode, \code{"overview"} or \code{"detailed"}}

\item{fix.col}{If \code{TRUE}, color distribution is fixed.}

\item{no.clst}{If \code{TRUE}, no clustering lines are plotted.}

\item{top}{If \code{TRUE}, use total abs(correlation) in selection.}

\item{top.num}{The number of the features with the best contrast to be plotted.}
}
\description{
Plot correlations heatmap and write an output file \code{"correlation_table.csv"}.
Note, the heatmap will only show correlations for a maximum of 1000 features. 
For larger datasets, only top 1000 features will be selected based on their interquantile range (IQR). 
When color distribution is fixed, you can potentially compare the correlation patterns among different data sets. 
In this case, you can choose "do not perform clustering" for all data set, 
or only to perform clustering on a single reference data set, 
then manually re-arranged other data sets according to the clustering pattern of the reference data set.
}
\seealso{
\code{\link{PatternHunter}}, \code{\link{PlotCorr}}
}

