% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_EBAM.R
\name{PlotEBAM.A0}
\alias{PlotEBAM.A0}
\alias{PlotEBAM.Cmpd}
\title{\code{PlotEBAM.A0} - plot EBAM A0 plot.}
\usage{
PlotEBAM.A0(dataSet, analSet, imgName = "ebam_view_", format = "png",
  dpi = 72, width = NA)

PlotEBAM.Cmpd(dataSet, analSet, imgName = "ebam_imp_", format = "png",
  dpi = 72, width = NA)
}
\description{
\code{PlotEBAM.A0} - plot EBAM A0 plot.

\code{PlotEDAM.Cmpd} - plot significant features.
}

