% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.R
\name{PlotFC}
\alias{PlotFC}
\title{Plot fold change}
\usage{
PlotFC(dataSet, analSet, imgName = "fc_", format = "png", dpi = 72,
  width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Plot fold change analysis results.
}
\seealso{
\code{\link{FC.Anal}} for analytical function
}

