% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dendrogram.R
\name{PlotHCTree}
\alias{PlotHCTree}
\title{Plot dendrogram}
\usage{
PlotHCTree(dataSet, analSet, imgName = "tree_", format = "png", dpi = 72,
  width = NA, smplDist = "euclidean", clstDist = "ward.D")
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{smplDist}{The distance measure. One of \code{"euclidean"}, \code{"spearman"}}

\item{clstDist}{The agglomeration method to be used, one of "ward.D", "ward.D2", "single", "complete", "average".
For details: \code{\link[stats]{hclust}}}
}
\description{
Plot hierarchical clustering tree.
}

