% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HeatMap.R
\name{PlotHeatMap2}
\alias{PlotHeatMap2}
\title{Plot two-factor heatmap}
\usage{
PlotHeatMap2(dataSet, analSet, imgName = "heatmap2_", format = "png",
  dpi = 72, width = NA, smplDist = "pearson", clstDist = "average",
  colors = "bwm", viewOpt = "overview", hiRes = FALSE, sortInx = 2,
  var.inx = 1:ncol(dataSet$norm))
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{smplDist}{The distance measure, one of \code{"euclidean"}, \code{"pearson"}, \code{"minkowski"}}

\item{clstDist}{The agglomeration method to be used, one of "ward.D", "ward.D2", "single", "complete", "average".
For details: \code{\link[pheatmap]{pheatmap}}}

\item{colors}{The color contrast. One of \code{"default"}, \code{"gbr"} (red/green), \code{"heat"}, \code{"topo"}, \code{"gray"}}

\item{viewOpt}{View mode, \code{"overview"} or \code{"detailed"}}

\item{hiRes}{If \code{TRUE}, then produces hi-resolution plot.}

\item{sortInx}{If \code{A}, samples are arranged by the first factor; if \code{B} - by the second one.}

\item{var.inx}{Vector of the numbers of plotted features.}
}
\value{
Native \code{analSet} with one added \code{$htmap2} element containing:
\itemize{
\item\code{$htmap2$dist.par} - equal \code{smplDist} argument value
\item\code{$htmap2$clust.par} - equal \code{clstDist} argument value
}
}
\description{
Uses \code{\link[pheatmap]{pheatmap}} function.
This method displays data in the form of colored cells.
It provides direct visualization of the relative levels of individual samples or variables.
Each colored cell on the map corresponds to a concentration value in your data table,
with samples in rows and features/compounds in columns.
You can use heatmap to identify samples/features that are unusually high/low.
}

