% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster_Kmeans.R
\name{PlotKmeans}
\alias{PlotKmeans}
\title{Plot K-means}
\usage{
PlotKmeans(dataSet, analSet, imgName = "kmeans_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Plot K-means map.
Please note: only the cluster members will be calculated
if the specified cluster number > 20.
The blue lines represent the median intensities of each cluster.
}

