% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA.R
\name{PlotPCA}
\alias{PlotPCA}
\alias{PlotPCA2DScore}
\alias{PlotPCABiplot}
\alias{PlotPCALoadings}
\alias{PlotPCAPairSummary}
\alias{PlotPCAScree}
\title{Plot PCA}
\usage{
PlotPCAPairSummary(dataSet, analSet, imgName = "pca_pair_", format = "png",
  dpi = 72, width = NA, pc.num = 2)

PlotPCAScree(dataSet, analSet, imgName = "pca_scree_", format = "png",
  dpi = 72, width = NA, scree.num)

PlotPCA2DScore(dataSet, analSet, imgName = "pca_score2d_", format = "png",
  dpi = 72, width = NA, inx1 = 1, inx2 = 2, reg = 0.95, show = TRUE,
  grey.scale = FALSE)

PlotPCABiplot(dataSet, analSet, imgName = "pca_biplot_", format = "png",
  dpi = 72, width = NA, inx1 = 1, inx2 = 2)

PlotPCALoadings(dataSet, analSet, imgName = "pca_loading_", format = "png",
  dpi = 72, width = NA, plotType = "scatter", show = TRUE)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{pc.num}{The number of plotted principal components.}

\item{scree.num}{The total number of plotted PCs.}

\item{inx1,inx2}{The order numbers of PCs.}

\item{reg}{Set the confidence level for plotting confidence region ellipse.}

\item{show}{If \code{TRUE} then points at the plot are labeled.}

\item{grey.scale}{If \code{TRUE} then plot is colored in 50 shades of gray.}

\item{plotType}{\code{"scatter"} for scatter plot or \code{"boxplot"} for box plot}

\item{inx1,inx2}{The order numbers of PCs.}

\item{show}{If \code{TRUE} then points at the plot are labeled.}
}
\description{
Set of functions for plotting the results of PCA.

\code{PlotPCAPairSummary} - plot summary.

\code{PlotPCAScree} - plot scree plot of varience explained by PC.

\code{PlotPCA2DScore} - plot 2D score plot.

\code{PlotPCABiplot} - plot the biplot.

\code{PlotPCALoadings} - plot loadings.
}

