% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSImp.R
\name{PlotPLS.Imp}
\alias{PlotPLS.Imp}
\title{Plot PLS important features}
\usage{
PlotPLS.Imp(dataSet, analSet, imgName = "pls_imp_", format = "png",
  dpi = 72, width = NA, type = "vip", feat.nm, feat.num = 15,
  color.BW = FALSE)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{type}{Importance measure. \code{"vip"} - for variable importance in projection (VIP); \code{"coef"} - for weighted sum of absolute regression coefficients.}

\item{feat.nm}{??????????????????????????????????????}

\item{feat.num}{The number of top features to plot.}

\item{color.BW}{If \code{TRUE}, grey scale is used.}
}
\description{
There are two importance measures in PLS-DA :
one is variable importance in projection (VIP)
and the other is weighted sum of absolute regression coefficients (coef.).
The colored boxes on the right indicate the relative concentrations
of the corresponding metabolite in each group under study.
}

