% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLS.R, R/PLSDA.CV.R, R/PLSDA.Permut.R
\name{PlotPLS}
\alias{PlotPLS}
\alias{PlotPLS.Classification}
\alias{PlotPLS.Permutation}
\alias{PlotPLS2DScore}
\alias{PlotPLSLoading}
\alias{PlotPLSPairSummary}
\title{Plot PLS}
\usage{
PlotPLSPairSummary(dataSet, analSet, imgName = "pls_pair_", format = "png",
  dpi = 72, width = NA, pc.num = 2)

PlotPLS2DScore(dataSet, analSet, imgName = "pls_score2d_", format = "png",
  dpi = 72, width = NA, inx1 = 1, inx2 = 2, reg = 0.95, show = TRUE,
  grey.scale = F)

PlotPLSLoading(dataSet, analSet, imgName = "pls_loading_", format = "png",
  dpi = 72, width = NA, plotType = "scatter", show = T)

PlotPLS.Classification(dataSet, analSet, imgName = "plsda_", format = "png",
  dpi = 72, width = NA)

PlotPLS.Permutation(dataSet, analSet, imgName = "pls_perm_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}

\item{pc.num}{The number of plotted principal components.}

\item{inx1,inx2}{The order numbers of PCs.}

\item{reg}{Set the confidence level for plotting confidence region ellipse.}

\item{show}{If \code{TRUE} then points at the plot are labeled.}

\item{grey.scale}{If \code{TRUE} then plot is colored in 50 shades of gray.}

\item{plotType}{\code{"scatter"} for scatter plot or \code{"boxplot"} for box plot}
}
\description{
Set of functions for plotting the results of PLS.

\code{PlotPLSPairSummary} - plot summary.

\code{PlotPLS2DScore} - plot 2D score plot.

\code{PlotPLSLoadings} - plot loadings.

\code{PlotPLS.Classification} - plot PLSDA classification performance using different components

\code{PlotPLS.Permutation} - plot PLSDA permutation plot.
}
\seealso{
\code{\link{PLS.Anal}}, \code{\link{PLS.Loadings}}, \code{\link{PLSDA.CV}}, \code{\link{PLSDA.Permut}}
}

