% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RF.R
\name{PlotRF}
\alias{PlotRF}
\alias{PlotRF.Classification}
\alias{PlotRF.Outlier}
\alias{PlotRF.VIP}
\title{Plot Random Forest}
\usage{
PlotRF.Classification(dataSet, analSet, imgName = "rf_cls_", format = "png",
  dpi = 72, width = NA)

PlotRF.VIP(dataSet, analSet, imgName = "rf_imp_", format = "png",
  dpi = 72, width = NA)

PlotRF.Outlier(dataSet, analSet, imgName = "rf_outiler_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Set of functions for plotting the results of Random Forest analysis.

\code{PlotRF.Classification} - plot feature classification.

\code{PlotRF.VIP} - plot variable importance of features ranked by their contributions to classification accuracy (MeanDecreaseAccuracy).

\code{PlotRF.Outlier} - plot outlying measures, only top 5 potential outliers are labeled.
}
\seealso{
\code{\link{RF.Anal}} for analytical function
}

