% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RSVM.R
\name{PlotRSVM}
\alias{PlotRSVM}
\alias{PlotRSVM.Classification}
\alias{PlotRSVM.Cmpd}
\title{Plot RSVM}
\usage{
PlotRSVM.Classification(dataSet, analSet, imgName = "svm_cls_",
  format = "png", dpi = 72, width = NA)

PlotRSVM.Cmpd(dataSet, analSet, imgName = "svm_imp_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Set of functions for plotting the results of recursive Support Vector Machine analysis.

\code{PlotRSVM.Classification} - plot feature classification.

\code{PlotRSVM.VIP} - please note : features are ranked by their frequencies being selected in the best classifiers (only top 15 will be shown)
}
\seealso{
\code{\link{RSVM.Anal}} for analytical function
}

