% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Cluster_SOM.R
\name{PlotSOM}
\alias{PlotSOM}
\title{Plot SOM}
\usage{
PlotSOM(dataSet, analSet, imgName = "som_", format = "png", dpi = 72,
  width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Plot SOM map for  less than 20 clusters.
Please note : only cluster members will be calculated if the total cluster number (xdim*ydim) > 20.
The blue lines represent the median intensities of each cluster.
}
\seealso{
\code{\link{SOM.Anal}} for analytical function
}

