% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Volcano.R
\name{PlotVolcano}
\alias{PlotVolcano}
\title{Plot Volcano results}
\usage{
PlotVolcano(dataSet, analSet, imgName = "volcano_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
The volcano plot is a combination of fold change and t-tests.
Note, for unpaired samples, the x-axis is log (FC).
For paired analysis, the x-axis is number of significant counts.
Y-axis is -log10(p.value) for both cases.
}
\seealso{
\code{\link{Volcano.Anal}} for analytical function
}

