% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLS.R
\name{PLS.Loadings}
\alias{PLS.Loadings}
\title{PLS loadings}
\usage{
PLS.Loadings(dataSet, analSet, inx1 = 1, inx2 = 2)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{inx1,inx2}{The numbers of PC.}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$pls$load.x.uniq} - ???
\item\code{$pls$imp.loads} - loading matrix
}
}
\description{
PLS loadings.
Adds loading matrix to \code{analSet$pls} element.
}
\seealso{
\code{\link{PLS.Anal}}, \code{\link{PLSDA.CV}}, \code{\link{PLSDA.Permut}}, \code{\link{PlotPLS}}
}

