% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSDA.CV.R
\name{PLSDA.CV}
\alias{PLSDA.CV}
\title{PLSDA cross validation}
\usage{
PLSDA.CV(dataSet, analSet, methodName = "CV",
  compNum = GetDefaultPLSCVComp(dataSet), choice = "Q2")
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{methodName}{Cross validation method \code{"LOOCV"} or \code{"CV"}}

\item{compNum}{The maximum number of components to search.}

\item{choice}{The determining performance measure, one of \code{"Q2"}, \code{"R2"}, \code{"Accuracy"}}
}
\value{
Native \code{analSet} with one added \code{$plsda} element consisting of:
\itemize{
\item\code{$best.num} - the detected best number of components
\item\code{$choice} - value of \code{choice} argument
\item\code{$coef.mat} - matrix of VIPs (variable importance in projection)
\item\code{$vip.mat} - matrix of weighted sums of absolute regression coefficients
\item\code{$fit.info} - all performance measures calculated for different number of components
}
}
\description{
Perform PLSDA classification and feature selection.
Writes two output files: \code{"plsda_vip.csv"} and \code{"plsda_coef.csv"}.
}
\seealso{
\code{\link{PLS.Anal}}, \code{\link{PLS.Loadings}}, \code{\link{PLSDA.Permut}}, \code{\link{PlotPLS}}
}

