% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLSDA.Permut.R
\name{PLSDA.Permut}
\alias{PLSDA.Permut}
\title{Permutation PLSDA test}
\usage{
PLSDA.Permut(dataSet, analSet, num = 100, type = "accu")
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{num}{The number of permutations.}

\item{type}{The indicator. If \code{"accu"} - training classification accuracy;
if \code{"sep"} - separation distance.}
}
\value{
Native \code{analSet} with some elements added:
\itemize{
\item\code{$plsda$permut.p} -
\item\code{$plsda$permut.inf} -
\item\code{$plsda$permut.type} -
\item\code{$plsda$permut} -
}
}
\description{
Perform permutation, using training classification accuracy or separation distance as
indicator, for two or multi-groups.
}

