% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RF.R
\name{RF.Anal}
\alias{RF.Anal}
\title{Random Forests}
\usage{
RF.Anal(dataSet, analSet, treeNum = 500, tryNum = 10)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{treeNum}{Number of trees to grow.}

\item{tryNum}{Number of predictors to try for each node.}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$rf} - standard \code{\link[randomForest]{randomForest}} function output
\item\code{$rf.sigmat} - matrix of features significance (sorted by MeanDecreaseAccuracy).
}
}
\description{
Perform Random Forest analysis. Uses \code{\link[randomForest]{randomForest}} function.
Writes results to \code{"randomforests_sigfeatures.csv"} file.
}

