% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RSVM.R
\name{RSVM.Anal}
\alias{RSVM.Anal}
\title{Support Vector Machine (SVM)}
\usage{
RSVM.Anal(dataSet, analSet, cvType = 10)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{cvType}{Type of cross-validation of: integer - N-fold-CV is used;
\code{"LOO"} - leave-one-out CV (LOOCV); \code{"bootstrape"} - bootstrape CV}
}
\value{
Native \code{analSet} with one added element \code{$svm} containing
- standard \code{\link[randomForest]{randomForest}} function output
}
\description{
Perform recursive SVM for feature selection and classification. Uses \code{\link[e1071]{svm}} function.
Writes results to \code{"svm_sigfeatures.csv"} file.
R-SVM uses SVM (with linear kernel) to perform classification recursively
using different feature subsets. Features are selected based on their relative
contribution in the classification using cross validation error rates.
The least important features are eliminated in the subsequent steps.
This process creates a series of SVM models (levels).
The features used by the best model are plotted.
LOOCV: leave one out cross-validation.
}
\seealso{
\code{\link[e1071]{svm}} for used statistical function\cr
\code{\link{PlotRSVM}} for plotting functions
}

