% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_SAM.R
\name{SAM.Anal}
\alias{SAM.Anal}
\title{SAM analysis}
\usage{
SAM.Anal(dataSet, analSet, method = "d.stat", paired = FALSE,
  var.equal = TRUE)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{method}{A character string or a name specifying the method/function that
should be used in the computation of the expression scores d.
If \code{"d.stat"}, a modified t-statistic or F-statistic, respectively, will be computed as proposed by Tusher et al. (2001).
If \code{"wilc.stat"}, a Wilcoxon rank sum statistic or Wilcoxon signed rank statistic will be used as expression score.}

\item{paired}{Are values in data set paired or not.}

\item{var.equal}{Are variances assumed equal or not.}
}
\value{
Native \code{analSet} with one added \code{$sam} element containing
standard \code{\link[siggenes]{sam}} function output
}
\description{
Perform Significance Analysis of Microarray (and Metabolites). Uses \code{\link[siggenes]{sam}} function.
}

