% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetColor_SetShape.R
\name{SetColor}
\alias{SetColor}
\title{Set color schema}
\usage{
SetColor(dataSet, colVec, msdata, fac)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{colVec}{The vector of color names, numbers or RGB encoded values}

\item{msdata}{\code{\link[MSdata]{MSdata-class}} object used for \code{dataSet} generation}

\item{fac}{The name of the factor corresponding to difference in shape.}
}
\description{
Set the color schema for plotting the results of analysis.
}
\seealso{
\code{\link[grDevices]{colors}}
}

