% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_EBAM.R
\name{SetEBAMSigMat}
\alias{SetEBAMSigMat}
\title{EBAM matrix of significance}
\usage{
SetEBAMSigMat(dataSet, analSet, delta = 0.9)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{delta}{The delta to control FDR}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$ebam.cmpd} - matrix containing \code{z.value}, \code{posterior}, \code{local.fdr} columns
\item\code{$ebam.delta} - value of \code{delta} argument
}
}
\description{
EBAM matrix of significance. Writes results to \code{"ebam_sigfeatures.csv"} file.
}

