% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetColor_SetShape.R
\name{SetShape}
\alias{SetShape}
\title{Set shape schema}
\usage{
SetShape(dataSet, shapeVec, msdata, fac)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{shapeVec}{The vector of point shapes (see corresponging numbers here: \code{\link[graphics]{points}}).}

\item{msdata}{\code{\link[MSdata]{MSdata-class}} object used for \code{dataSet} generation}

\item{fac}{The name of the factor corresponding to difference in shape.}
}
\description{
Set the shape schema for plotting the results of analysis.
}
\seealso{
\code{\link[graphics]{points}}
}

