% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttests.R
\name{Ttests.Anal}
\alias{Ttests.Anal}
\title{T-test}
\usage{
Ttests.Anal(dataSet, analSet, nonpar = FALSE, thresh = 0.05,
  paired = FALSE, var.equal = TRUE)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{nonpar}{If \code{FALSE} - use classical t-test; if \code{TRUE} - Wilcoxon Rank Test}

\item{thresh}{Threshold of significance.}

\item{paired}{Are values in data set paired or not.}

\item{var.equal}{Are variances assumed equal or not.}
}
\value{
Native \code{analSet} with one added \code{$tt} element consisting of:
\itemize{
\item\code{$tt.nm} - method of analysis 
\item\code{$raw.thresh} - value of \code{thresh} argument
\item\code{$thresh} - \code{-log10(thresh)}
\item\code{$p.value} - p value
\item\code{$p.log} - \code{-log10(p.value)}
\item\code{$inx.imp} - logical vector of features with significant difference
\item\code{$sig.mat} - data frame of significant features
}
}
\description{
Performs group comparison. Creates a file with results: 
either \code{"Wilcoxon_rank.csv"} or \code{"t_test.csv"}
depending on \code{nonpar} argument.
}
\note{
Note, for large data set (> 1000 variables), 
both the paired information and the group variance will be ignored, 
and the default parameters will be used for t-tests to save computational time. 
If you choose non-parametric tests (Wilcoxon rank-sum test), 
the group variance will be ignored.
}
\seealso{
\code{\link{PlotTT}} for plotting functions
}

