% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Volcano.R
\name{Volcano.Anal}
\alias{Volcano.Anal}
\title{Volcano}
\usage{
Volcano.Anal(dataSet, analSet, paired = FALSE, fcthresh = 2, cmpType = 0,
  percent.thresh = 0.75, nonpar = FALSE, thresh = 0.05,
  var.equal = TRUE)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{paired}{Are values in data set paired or not.}

\item{fcthresh}{Fold-change threshold.}

\item{cmpType}{Comparison type. If equal \code{0} then group 1 is compared against group 2. Otherwise, vice versa.}

\item{percent.thresh}{Sig. count threshold (for paired data).}

\item{nonpar}{If \code{FALSE} - use classical t-test; if \code{TRUE} - Wilcoxon Rank Test}

\item{thresh}{Threshold of significance.}

\item{var.equal}{Are variances assumed equal or not.}
}
\value{
Native \code{analSet} with one added \code{$volcano} element consisting of:
\itemize{
\item\code{$raw.threshx} - value of \code{fcthresh} argument
\item\code{$raw.threshy} - value of \code{thresh} argument
\item\code{$paired} - \code{-log10(thresh)}
\item\code{$max.xthresh} - upper log-threshold of fold change
\item\code{$min.xthresh} - lower log-threshold of fold change
\item\code{$thresh.y} - \code{-log10(thresh)}
\item\code{$fc.all} - fold changes of all features
\item\code{$fc.log} - log of fold changes
\item\code{$fc.log.uniq} - java-object
\item\code{$inx.up} - logical vector of increasing features
\item\code{$inx.down} - logical vector of decreasing features
\item\code{$p.log} - \code{-log10(p.value)}
\item\code{$inx.p} - logical vector of features with \code{p <= thresh}
\item\code{$sig.mat} - data frame of significant features
}
}
\description{
The volcano plot is a combination of fold change and t-tests.
Note, for unpaired samples, the x-axis is log (FC).
For paired analysis, the x-axis is number of significant counts.
Y-axis is -log10(p.value) for both cases.
Writes an output file \code{"volcano.csv"}
}

