% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA.Anal}
\alias{PCA.Anal}
\title{Perform PCA}
\usage{
PCA.Anal(dataSet, analSet)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}
}
\value{
Native \code{analSet} with one added \code{$pca} element consisting of:
\itemize{
\item all the elements of the standard \code{\link[stats]{prcomp}} output
\item\code{$std} - standard deviation
\item\code{$variance} - variance explained by each PC
\item\code{$cum.var} - cummulated variance explained
}
}
\description{
Uses \code{\link[stats]{prcomp}} function.
Writes two output files: \code{"pca_score.csv"} and \code{"pca_loadings.csv"}.
Adds \code{analSet$pca} element with \code{\link[stats]{prcomp}} function output
and some basic statistics.
}

