% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_EBAM.R
\name{PlotEBAM}
\alias{PlotEBAM}
\alias{PlotEBAM.A0}
\alias{PlotEBAM.Cmpd}
\title{Plot EBAM}
\usage{
PlotEBAM.A0(dataSet, analSet, imgName = "ebam_view_", format = "png",
  dpi = 72, width = NA)

PlotEBAM.Cmpd(dataSet, analSet, imgName = "ebam_imp_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Functions for plotting the results of EBAM.

\code{PlotEBAM.A0} - plot EBAM A0 plot.

\code{PlotEDAM.Cmpd} - plot significant features.
}

