% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Signfeat_SAM.R
\name{PlotSAM}
\alias{PlotSAM}
\alias{PlotSAM.Cmpd}
\alias{PlotSAM.FDR}
\title{Plot SAM}
\usage{
PlotSAM.FDR(dataSet, analSet, delta = NULL, imgName = "sam_view_",
  format = "png", dpi = 72, width = NA)

PlotSAM.Cmpd(dataSet, analSet, imgName = "sam_imp_", format = "png",
  dpi = 72, width = NA)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}

\item{delta}{The delta to control FDR. If \code{NULL}, then it's suggested automaticly}

\item{imgName}{Image file name prefix.}

\item{format}{Image format, one of: "png", "tiff", "pdf", "ps", "svg"}

\item{dpi}{Image resolution.}

\item{width}{Image width.}
}
\description{
Functions for plotting the results of SAM.

\code{PlotSAM.FDR} - plot FDR.

\code{PlotSAM.Cmpd} - plot significant features.
}

