% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLS.R
\name{PLS.Anal}
\alias{PLS.Anal}
\title{PLS analysis
Uses \code{\link[pls]{'oscorespls'}} method of \code{\link[pls]{plsr}} function .
Writes two output files: \code{"pls_score.csv"} and \code{"pls_loadings.csv"}.
Adds \code{analSet$pls} element with \code{\link[pls]{plsr}} function output.}
\usage{
PLS.Anal(dataSet, analSet)
}
\arguments{
\item{dataSet}{}

\item{analSet}{}
}
\value{
Native \code{analSet} with one added \code{$plsr} element containing \code{\link[pls]{plsr}} function output
}
\description{
PLS analysis
Uses \code{\link[pls]{'oscorespls'}} method of \code{\link[pls]{plsr}} function .
Writes two output files: \code{"pls_score.csv"} and \code{"pls_loadings.csv"}.
Adds \code{analSet$pls} element with \code{\link[pls]{plsr}} function output.
}

