% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PLS.R
\name{PLS.Loadings}
\alias{PLS.Loadings}
\title{PLS loadings}
\usage{
PLS.Loadings(dataSet, analSet, imgName, format = "png", dpi = 72,
  width = NA, inx1 = 1, inx2 = 2, plotType, lbl.feat = 1)
}
\arguments{
\item{analSet}{}

\item{inx1,inx2}{The numbers of PC.}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$pls$load.x.uniq} - ???
\item\code{$pls$imp.loads} - loading matrix
}
}
\description{
PLS loadings.
Adds loading matrix to \code{analSet$pls} element.
}

