% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signfeat_SAM.R
\name{SetSAMSigMat}
\alias{SetSAMSigMat}
\title{SAM matrix of significance}
\usage{
SetSAMSigMat(dataSet, analSet, delta = 1.3)
}
\arguments{
\item{dataSet}{List, data set object generated by \code{\link[MSdata]{MS_to_MA}} function.}

\item{analSet}{List, containing the results of statistical analysis (can be just an empty list).}

\item{delta}{The delta to control FDR}
}
\value{
Native \code{analSet} with two added elements:
\itemize{
\item\code{$sam.cmpd} - matrix containing \code{d.value}, \code{stdev}, \code{rawp}, \code{q.value} columns
\item\code{$sam.delta} - value of \code{delta} argument
}
}
\description{
SAM matrix of significance. Writes results to \code{"sam_sigfeatures.csv"} file.
}
\seealso{
\code{\link{SAM.Anal}} for analytical function\cr
\code{\link{PlotSAM}} for plotting functions
}

